move mock puppet to a new module named wechaty-puppet-mock
