+++ pom.xml
@@ -32,6 +32,7 @@
         <module>wechaty</module>
         <module>wechaty-puppet</module>
         <module>wechaty-puppet-hostie</module>
+        <module>wechaty-puppet-mock</module>
     </modules>
 
     <dependencyManagement>
@@ -187,13 +188,11 @@
                 <groupId>org.mockito</groupId>
                 <artifactId>mockito-core</artifactId>
                 <version>3.3.3</version>
-                <scope>test</scope>
             </dependency>
             <dependency>
                 <groupId>junit</groupId>
                 <artifactId>junit</artifactId>
                 <version>4.12</version>
-                <scope>test</scope>
             </dependency>
 
         </dependencies>
