+++ wechaty-puppet-mock/src/main/resources/mockito-extensions/org.mockito.plugins.MockMaker
